Servico de items

+++ dasa-receituario-dao/src/main/java/br/com/tamanhofamilia/dasa/receituario/daos/pedidos/PedidoItemDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PedidoItemDao extends JpaRepository<PedidoItem, Integer> {
+public interface PedidoItemDao extends JpaRepository<PedidoItem, Long> {
 }
